agregate Response Status to BandMemberController and ArtistController

+++ src/main/java/ar/edu/unrc/exa/dc/dose2016/riocuartobandasderock/main/ArtistController.java
@@ -1,6 +1,7 @@
 package ar.edu.unrc.exa.dc.dose2016.riocuartobandasderock.main;
 
 import ar.edu.unrc.exa.dc.dose2016.riocuartobandasderock.model.Artist;
+import ar.edu.unrc.exa.dc.dose2016.riocuartobandasderock.model.BandMember;
 import spark.Response;
 import spark.Request;
 import ar.edu.unrc.exa.dc.dose2016.riocuartobandasderock.dao.ArtistDAO;
@@ -16,13 +17,6 @@ public class ArtistController {
 	 */
 	private ArtistDAO artistDAO;
 	
-	/**
-	 * 
-	 * @return ArtistDAO;
-	 */
-	public ArtistDAO getArtistDAO() {
-		return artistDAO;
-	}
     /**
      * Constructor
      */
@@ -37,7 +31,12 @@ public class ArtistController {
 	 * @return  List of Artist
 	 */
 	public List<Artist> getAllArtists(Request req, Response res){
-		return artistDAO.getAllArtists(); 
+		//artistDAO.openCurrentSession();
+		List<Artist> artists = artistDAO.getAllArtists();
+		//artistDAO.closeCurrentSession();
+		int status = (artists.size()>0)? 200:404;
+		res.status(status);
+		return artists;
 	}
 	
 	/**
@@ -48,8 +47,15 @@ public class ArtistController {
 	 */
 	
 	public Artist getArtistById (Request req, Response res){
-		if (req==null) throw new IllegalArgumentException("Reques is null");
-		return artistDAO.findById(req.params(":id")); 
+		if (req.params(":id")==""){
+			res.status(400);
+		}
+		//artistDAO.openCurrentSession();
+		Artist artist = artistDAO.findById(req.params(":id"));
+		//artistDAO.closeCurrentSession();
+		int status = (artist==null)? 201:404;
+		res.status(status);
+		return artist;
 	}
 	
 	/**
@@ -59,49 +65,92 @@ public class ArtistController {
 	 * @return List artist with name parameters
 	 */
 	public List<Artist> getArtistByName (Request req, Response res){
-		return artistDAO.findByName(req.params("name"));
+		if (req.params(":name")==""){
+			res.status(400);
+		}
+		//artistDAO.openCurrentSession();
+		List<Artist> artists = artistDAO.findByName(req.params(":name"));
+		//artistDAO.closeCurrentSession();
+		int status = (artists.size()==0)? 201:404;
+		res.status(status);
+		return artists;
 	}
 		
 	/**
 	 * creates an artist 
 	 * @param req    It contains the attributes of the new artist
 	 * @param res
-	 * @return True  when Artist It was created successfully
+	 * @return a string that describes the result of createArtist
 	 */
-	public boolean createArtist(Request req,Response res){
+	public String createArtist(Request req,Response res){
+		if((req.queryParams("name")=="") && (req.queryParams("surname")=="") && (req.queryParams("nickname")=="")){
+			res.status(400);
+			return "Request invalid";
+		}
 		Artist artist = new Artist(req.queryParams("name"),req.queryParams("surname"),req.queryParams("nickname"));
-		res.status(201);
+		//artistDAO.openCurrentSessionWithTransaction();
 		boolean status = artistDAO.createArtist(artist);
-		return status;
+		//artistDAO.closeCurrentSessionWithTransaction();
+		if (status){
+			res.status(201);
+			return "Success";
+		}
+		res.status(500);
+		return "Fail";	
 	}
 	
 	/**
 	 * update an artist 
 	 * @param req	It contains the attributes of the artist to update
 	 * @param res
-	 * @return True when artist successfully updated
+	 * @return a string that describes the result of updateBandMember
 	 */
-	public boolean updateArtist(Request req, Response res){
+	public String updateArtist(Request req, Response res){
+		if((req.queryParams("name")=="") && (req.queryParams("surname")=="") && (req.queryParams("nickname")=="")){
+			res.status(400);
+			return "Request invalid";
+		}
+		//artistDAO.openCurrentSession();
 		Artist artist = artistDAO.findById(req.params(":id"));
+		//artistDAO.closeCurrentSession();
 		if (artist==null){
-			return false;
+			res.status(400);
+			return "Request invalid";
 		}
 		artist.setName(req.queryParams("name"));
 		artist.setNickname(req.queryParams("nickName"));
 		artist.setSurname(req.queryParams("surName"));
+		//artistDAO.openCurrentSessionWithTransaction();
 		boolean status = artistDAO.updateArtist(artist);
-		return status;
+		//artistDAO.closeCurrentSessionWithTransaction();
+		if (status){
+			res.status(200);
+			return "Success";
+		}
+		res.status(500);
+		return "Fail";
 	}
 	
 	/**
 	 * delete an artist by his Id
 	 * @param req it contain id of the artist to delete
 	 * @param res
-	 * @return True when Artist successfully deleted
+	 * @return a string that describes the result of deleteBandMember
 	 */
-	public boolean deleteArtist(Request req, Response res){
+	public String deleteArtist(Request req, Response res){
+		if ((req.params(":id"))==""){
+			res.status();
+			return "Request invalid";
+		}
+		//artistDAO.openCurrentSessionWithTransaction();
 	 	boolean status = artistDAO.deleteArtist(req.params(":id"));
-	 	return status;
+		//artistDAO.closeCurrentSessionWithTransaction();
+		if (status){
+			res.status(200);
+			return "Success";
+		}
+		res.status(404);
+		return "Fail";
 	}
 
 }
